detector/requirements.txt
fastapi
easyocr
uvicorn
pandas
numpy
python-multipart
Pillow
opencv-python==4.5.5.64
